Transition Shooting to Trace
Removed all the Projectile logic and replaced it with a line trace.
I kept in the behavior as before to be consistent.
I removed all the projectiles for it too just to reduce the size also as it is not needed, so we don't want it.
If we ever want to go back, there's history in git to revert or take a basis for reimplementation.