Enable Multiplayer
Enable all the Multiplayer Functionality and make everything RPC properly.
I decided to set the owner of the weapon, where all the input is handled, to the clients, mostly due to the input handling. This could cause the player to cheat, as we also handle ammo within the weapon. But it was not a requirement, I'm still pointing it out.
I've also added some debug infos as networking and replication is always difficult to debug, those are not really saying much (I do prefer the blueprint prints which also show the caller of it), but they do their job in telling what is called when.